[FLINK-40106] Move FlinkResourceValidator and FlinkResourceMutator to the api module
