+++ flink-kubernetes-operator/src/main/java/org/apache/flink/kubernetes/operator/mutator/DefaultFlinkMutator.java
@@ -21,6 +21,7 @@ import org.apache.flink.kubernetes.operator.api.CrdConstants;
 import org.apache.flink.kubernetes.operator.api.FlinkDeployment;
 import org.apache.flink.kubernetes.operator.api.FlinkSessionJob;
 import org.apache.flink.kubernetes.operator.api.FlinkStateSnapshot;
+import org.apache.flink.kubernetes.operator.api.mutator.FlinkResourceMutator;
 
 import java.util.HashMap;
 import java.util.Optional;
