+++ flink-kubernetes-operator/src/test/java/org/apache/flink/kubernetes/operator/validation/TestValidator.java
@@ -21,6 +21,7 @@ import org.apache.flink.kubernetes.operator.api.AbstractFlinkResource;
 import org.apache.flink.kubernetes.operator.api.FlinkDeployment;
 import org.apache.flink.kubernetes.operator.api.FlinkSessionJob;
 import org.apache.flink.kubernetes.operator.api.FlinkStateSnapshot;
+import org.apache.flink.kubernetes.operator.api.validation.FlinkResourceValidator;
 
 import java.util.Optional;
